tmux: Fix copy behaviour
Removing the `-and-cancel` part was a bad idea - this leads to the paste
buffer just being constantly appended to, rather than replaced, when
copying a string.

I'm sure there's a solution to this, e.g. explicitly clearing the buffer
before running `pbcopy`, but for now I'd prefer it works and am not too
fussed about being able to do multiple copies within the same
scrollback.